email_forwarder/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiohttp==3.9.1
pydantic==2.5.2
email-validator==2.1.0
python-multipart==0.0.6 